requirements.txt
pymodbus==3.5.4
paho-mqtt==1.6.1